services/match_details/pyproject.toml
[tool.poetry]
name = "Lightshield_Match_Details_Manager"
version = "2.0.0"
description = "Base image for Lightshield services."
authors = ["Doctress"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.9"
aioredis = "^1.3.1"
asyncpg = "^0.23.0"
aiohttp = "^3.7.4"
uvloop = "^0.15.2"
lightshield = '*'
guppy3 = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
